requirements.txt
PyGithub>=1.55
python-dotenv>=0.19.0
fpdf==1.7.2
